+++ scripts/codegen/03-structs.go
@@ -1,39 +1,48 @@
 package main
 
 type LocaleData struct {
-	Name              string                `json:"name,omitempty"`
-	DateOrder         string                `json:"date_order,omitempty"`
-	January           []string              `json:"january,omitempty"`
-	February          []string              `json:"february,omitempty"`
-	March             []string              `json:"march,omitempty"`
-	April             []string              `json:"april,omitempty"`
-	May               []string              `json:"may,omitempty"`
-	June              []string              `json:"june,omitempty"`
-	July              []string              `json:"july,omitempty"`
-	August            []string              `json:"august,omitempty"`
-	September         []string              `json:"september,omitempty"`
-	October           []string              `json:"october,omitempty"`
-	November          []string              `json:"november,omitempty"`
-	December          []string              `json:"december,omitempty"`
-	Monday            []string              `json:"monday,omitempty"`
-	Tuesday           []string              `json:"tuesday,omitempty"`
-	Wednesday         []string              `json:"wednesday,omitempty"`
-	Thursday          []string              `json:"thursday,omitempty"`
-	Friday            []string              `json:"friday,omitempty"`
-	Saturday          []string              `json:"saturday,omitempty"`
-	Sunday            []string              `json:"sunday,omitempty"`
-	AM                []string              `json:"am,omitempty"`
-	PM                []string              `json:"pm,omitempty"`
-	Year              []string              `json:"year,omitempty"`
-	Month             []string              `json:"month,omitempty"`
-	Week              []string              `json:"week,omitempty"`
-	Day               []string              `json:"day,omitempty"`
-	Hour              []string              `json:"hour,omitempty"`
-	Minute            []string              `json:"minute,omitempty"`
-	Second            []string              `json:"second,omitempty"`
-	RelativeType      map[string][]string   `json:"relative-type,omitempty"`
-	RelativeTypeRegex map[string][]string   `json:"relative-type-regex,omitempty"`
-	LocaleSpecific    map[string]LocaleData `json:"locale_specific,omitempty"`
+	Name                  string                `json:"name,omitempty"                    yaml:"name,omitempty"`
+	DateOrder             string                `json:"date_order,omitempty"              yaml:"date_order,omitempty"`
+	SkipWords             []string              `json:"skip,omitempty"                    yaml:"skip,omitempty"`
+	PertainWords          []string              `json:"pertain,omitempty"                 yaml:"pertain,omitempty"`
+	NoWordSpacing         bool                  `json:"no_word_spacing,omitempty"         yaml:"no_word_spacing,omitempty"`
+	SentenceSplitterGroup int                   `json:"sentence_splitter_group,omitempty" yaml:"sentence_splitter_group,omitempty"`
+	January               []string              `json:"january,omitempty"                 yaml:"january,omitempty"`
+	February              []string              `json:"february,omitempty"                yaml:"february,omitempty"`
+	March                 []string              `json:"march,omitempty"                   yaml:"march,omitempty"`
+	April                 []string              `json:"april,omitempty"                   yaml:"april,omitempty"`
+	May                   []string              `json:"may,omitempty"                     yaml:"may,omitempty"`
+	June                  []string              `json:"june,omitempty"                    yaml:"june,omitempty"`
+	July                  []string              `json:"july,omitempty"                    yaml:"july,omitempty"`
+	August                []string              `json:"august,omitempty"                  yaml:"august,omitempty"`
+	September             []string              `json:"september,omitempty"               yaml:"september,omitempty"`
+	October               []string              `json:"october,omitempty"                 yaml:"october,omitempty"`
+	November              []string              `json:"november,omitempty"                yaml:"november,omitempty"`
+	December              []string              `json:"december,omitempty"                yaml:"december,omitempty"`
+	Monday                []string              `json:"monday,omitempty"                  yaml:"monday,omitempty"`
+	Tuesday               []string              `json:"tuesday,omitempty"                 yaml:"tuesday,omitempty"`
+	Wednesday             []string              `json:"wednesday,omitempty"               yaml:"wednesday,omitempty"`
+	Thursday              []string              `json:"thursday,omitempty"                yaml:"thursday,omitempty"`
+	Friday                []string              `json:"friday,omitempty"                  yaml:"friday,omitempty"`
+	Saturday              []string              `json:"saturday,omitempty"                yaml:"saturday,omitempty"`
+	Sunday                []string              `json:"sunday,omitempty"                  yaml:"sunday,omitempty"`
+	AM                    []string              `json:"am,omitempty"                      yaml:"am,omitempty"`
+	PM                    []string              `json:"pm,omitempty"                      yaml:"pm,omitempty"`
+	Decade                []string              `json:"decade,omitempty"                  yaml:"decade,omitempty"`
+	Year                  []string              `json:"year,omitempty"                    yaml:"year,omitempty"`
+	Month                 []string              `json:"month,omitempty"                   yaml:"month,omitempty"`
+	Week                  []string              `json:"week,omitempty"                    yaml:"week,omitempty"`
+	Day                   []string              `json:"day,omitempty"                     yaml:"day,omitempty"`
+	Hour                  []string              `json:"hour,omitempty"                    yaml:"hour,omitempty"`
+	Minute                []string              `json:"minute,omitempty"                  yaml:"minute,omitempty"`
+	Second                []string              `json:"second,omitempty"                  yaml:"second,omitempty"`
+	In                    []string              `json:"in,omitempty"                      yaml:"in,omitempty"`
+	Ago                   []string              `json:"ago,omitempty"                     yaml:"ago,omitempty"`
+	RelativeType          map[string][]string   `json:"relative-type,omitempty"           yaml:"relative-type,omitempty"`
+	RelativeTypeRegex     map[string][]string   `json:"relative-type-regex,omitempty"     yaml:"relative-type-regex,omitempty"`
+	SimplificationList    []map[string]string   `json:"-"                                 yaml:"simplifications,omitempty"`
+	Simplifications       map[string]string     `json:"simplifications,omitempty"         yaml:"-"`
+	LocaleSpecific        map[string]LocaleData `json:"locale_specific,omitempty"         yaml:"locale_specific,omitempty"`
 }
 
 type CldrTerritoryData struct {