+++ scripts/codegen/data-cldr-parser.go
@@ -1,10 +1,8 @@
 package main
 
 import (
-	"encoding/json"
 	"os"
 	"path/filepath"
-	"sort"
 	"strconv"
 	"strings"
 )
@@ -13,12 +11,14 @@ func parseAllCldrData(languageLocalesMap map[string][]string) (map[string]Locale
 	result := map[string]LocaleData{}
 
 	for language, locales := range languageLocalesMap {
-		log.Info().Msgf("parsing language %s", language)
 		localeData, err := parseCldrData(language, locales...)
-		if err != nil {
+		if os.IsNotExist(err) {
+			continue
+		} else if err != nil {
 			return nil, err
 		}
 
+		log.Info().Msgf("parsed cldr %s", language)
 		result[language] = *localeData
 	}
 
@@ -37,10 +37,6 @@ func parseCldrData(locale string, sublocales ...string) (*LocaleData, error) {
 		return nil, err
 	}
 
-	if cldrGregorian == nil || cldrDateFields == nil {
-		return &LocaleData{}, nil
-	}
-
 	// Fetch main data
 	gregorianData := cldrGregorian.Main[locale].Dates.Calendars.Gregorian
 	dateFieldsData := cldrDateFields.Main[locale].Dates.Fields
@@ -48,7 +44,7 @@ func parseCldrData(locale string, sublocales ...string) (*LocaleData, error) {
 	// Generate locale data
 	setLocaleMonthData := func(month int) []string {
 		strMonth := strconv.Itoa(month)
-		return clearList(nil, localeMonthFilter,
+		return cleanList(nil, localeMonthFilter,
 			gregorianData.Months.Format.Wide[strMonth],
 			gregorianData.Months.Format.Abbreviated[strMonth],
 			gregorianData.Months.StandAlone.Wide[strMonth],
@@ -56,7 +52,7 @@ func parseCldrData(locale string, sublocales ...string) (*LocaleData, error) {
 	}
 
 	setLocaleWeekDays := func(weekDays string) []string {
-		return clearList(nil, nil,
+		return cleanList(nil, nil,
 			gregorianData.Days.Format.Wide[weekDays],
 			gregorianData.Days.Format.Abbreviated[weekDays],
 			gregorianData.Days.StandAlone.Wide[weekDays],
@@ -75,7 +71,7 @@ func parseCldrData(locale string, sublocales ...string) (*LocaleData, error) {
 			periodList[i] = rxAmPmPattern.ReplaceAllString(p, period)
 		}
 
-		return clearList(nil, nil, periodList...)
+		return cleanList(nil, nil, periodList...)
 	}
 
 	setLocaleDateFields := func(fieldName string, attr string) []string {
@@ -99,7 +95,7 @@ func parseCldrData(locale string, sublocales ...string) (*LocaleData, error) {
 			}
 		}
 
-		return clearList(nil, nil, finalList...)
+		return cleanList(nil, nil, finalList...)
 	}
 
 	setLocaleRegexes := func(fieldName string, relativeType string) []string {
@@ -124,7 +120,7 @@ func parseCldrData(locale string, sublocales ...string) (*LocaleData, error) {
 			}
 		}
 
-		return clearList(nil, regexFilter, finalList...)
+		return cleanList(nil, regexFilter, finalList...)
 	}
 
 	data := LocaleData{
@@ -158,7 +154,7 @@ func parseCldrData(locale string, sublocales ...string) (*LocaleData, error) {
 		Hour:      setLocaleDateFields("hour", "display"),
 		Minute:    setLocaleDateFields("minute", "display"),
 		Second:    setLocaleDateFields("second", "display"),
-		RelativeType: clearEmptyStringListMap(map[string][]string{
+		RelativeType: cleanEmptyStringListMap(map[string][]string{
 			"1 year ago":   setLocaleDateFields("year", "relative-past"),
 			"0 year ago":   setLocaleDateFields("year", "relative-now"),
 			"in 1 year":    setLocaleDateFields("year", "relative-future"),
@@ -175,7 +171,7 @@ func parseCldrData(locale string, sublocales ...string) (*LocaleData, error) {
 			"0 minute ago": setLocaleDateFields("minute", "relative-now"),
 			"0 second ago": setLocaleDateFields("second", "relative-now"),
 		}),
-		RelativeTypeRegex: clearEmptyStringListMap(map[string][]string{
+		RelativeTypeRegex: cleanEmptyStringListMap(map[string][]string{
 			"in \\1 year":    setLocaleRegexes("year", "future"),
 			"\\1 year ago":   setLocaleRegexes("year", "past"),
 			"in \\1 month":   setLocaleRegexes("month", "future"),
@@ -212,24 +208,12 @@ func parseCldrData(locale string, sublocales ...string) (*LocaleData, error) {
 }
 
 func parseCldrGregorianData(locale string) (*CldrGregorianData, error) {
-	// Prepare file path
+	// Parse JSON
 	cldrDatesFullDir := filepath.Join(RAW_DIR, "cldr-dates-full", "main")
 	fPath := filepath.Join(cldrDatesFullDir, locale, "ca-gregorian.json")
 
-	// Open file
-	f, err := os.Open(fPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			log.Warn().Msgf("gregorian data for locale %s not exist, skipped", locale)
-			return nil, nil
-		}
-		return nil, err
-	}
-	defer f.Close()
-
-	// Parse JSON
 	var data CldrGregorianData
-	err = json.NewDecoder(f).Decode(&data)
+	err := parseJsonFile(&data, fPath)
 	if err != nil {
 		return nil, err
 	}
@@ -251,157 +235,16 @@ func parseCldrGregorianData(locale string) (*CldrGregorianData, error) {
 }
 
 func parseCldrDateFieldsData(locale string) (*CldrDateFieldsData, error) {
-	// Prepare file path
 	cldrDatesFullDir := filepath.Join(RAW_DIR, "cldr-dates-full", "main")
 	fPath := filepath.Join(cldrDatesFullDir, locale, "dateFields.json")
 
-	// Open file
-	f, err := os.Open(fPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			log.Warn().Msgf("date fields data for locale %s not exist, skipped", locale)
-			return nil, nil
-		}
-		return nil, err
-	}
-	defer f.Close()
-
-	// Parse JSON
 	var data CldrDateFieldsData
-	err = json.NewDecoder(f).Decode(&data)
+	err := parseJsonFile(&data, fPath)
 	if err != nil {
 		return nil, err
 	}
 
 	return &data, nil
-}
-
-func cleanString(str string) string {
-	str = rxSanitizeAposthrope.ReplaceAllString(str, "'")
-	str = strings.ReplaceAll(str, ".", "")
-	str = strings.ToLower(str)
-	str = strings.Join(strings.Fields(str), " ")
-	return str
-}
-
-func clearList(initialItems []string, fnFilter func(string) bool, items ...string) []string {
-	// Prepare tracker
-	tracker := map[string]struct{}{}
-	if initialItems != nil && len(initialItems) > 0 {
-		for _, item := range initialItems {
-			tracker[item] = struct{}{}
-		}
-	}
-
-	// Create final list
-	var finalList []string
-	for _, item := range items {
-		item = cleanString(item)
-
-		// Make sure item not empty
-		if item == "" {
-			continue
-		}
-
-		// Make sure item tracked yet
-		if _, exist := tracker[item]; exist {
-			continue
-		}
-
-		// Make sure item passed the filter
-		if fnFilter != nil && !fnFilter(item) {
-			continue
-		}
-
-		// Save the item
-		tracker[item] = struct{}{}
-		finalList = append(finalList, item)
-	}
-
-	// Sort final list
-	sort.Strings(finalList)
-
-	if len(finalList) == 0 {
-		return nil
-	} else {
-		return finalList
-	}
-}
-
-func clearStringListMap(parentMap, childMap map[string][]string) map[string][]string {
-	for key, list := range childMap {
-		parentList := parentMap[key]
-		cleanedList := clearList(parentList, nil, list...)
-
-		if len(cleanedList) == 0 {
-			delete(childMap, key)
-		} else {
-			childMap[key] = cleanedList
-		}
-	}
-
-	if len(childMap) == 0 {
-		return nil
-	} else {
-		return childMap
-	}
-}
-
-func removeLocaleDataDuplicate(parent LocaleData, child LocaleData) LocaleData {
-	var cleanedDateOrder string
-	if child.DateOrder != parent.DateOrder {
-		cleanedDateOrder = child.DateOrder
-	}
-
-	return LocaleData{
-		Name:      child.Name,
-		DateOrder: cleanedDateOrder,
-		January:   clearList(parent.January, nil, child.January...),
-		February:  clearList(parent.February, nil, child.February...),
-		March:     clearList(parent.March, nil, child.March...),
-		April:     clearList(parent.April, nil, child.April...),
-		May:       clearList(parent.May, nil, child.May...),
-		June:      clearList(parent.June, nil, child.June...),
-		July:      clearList(parent.July, nil, child.July...),
-		August:    clearList(parent.August, nil, child.August...),
-		September: clearList(parent.September, nil, child.September...),
-		October:   clearList(parent.October, nil, child.October...),
-		November:  clearList(parent.November, nil, child.November...),
-		December:  clearList(parent.December, nil, child.December...),
-		Monday:    clearList(parent.Monday, nil, child.Monday...),
-		Tuesday:   clearList(parent.Tuesday, nil, child.Tuesday...),
-		Wednesday: clearList(parent.Wednesday, nil, child.Wednesday...),
-		Thursday:  clearList(parent.Thursday, nil, child.Thursday...),
-		Friday:    clearList(parent.Friday, nil, child.Friday...),
-		Saturday:  clearList(parent.Saturday, nil, child.Saturday...),
-		Sunday:    clearList(parent.Sunday, nil, child.Sunday...),
-		AM:        clearList(parent.AM, nil, child.AM...),
-		PM:        clearList(parent.PM, nil, child.PM...),
-		Year:      clearList(parent.Year, nil, child.Year...),
-		Month:     clearList(parent.Month, nil, child.Month...),
-		Week:      clearList(parent.Week, nil, child.Week...),
-		Day:       clearList(parent.Day, nil, child.Day...),
-		Hour:      clearList(parent.Hour, nil, child.Hour...),
-		Minute:    clearList(parent.Minute, nil, child.Minute...),
-		Second:    clearList(parent.Second, nil, child.Second...),
-
-		RelativeType:      clearStringListMap(parent.RelativeType, child.RelativeType),
-		RelativeTypeRegex: clearStringListMap(parent.RelativeTypeRegex, child.RelativeTypeRegex),
-	}
-}
-
-func clearEmptyStringListMap(m map[string][]string) map[string][]string {
-	for key, list := range m {
-		if len(list) == 0 {
-			delete(m, key)
-		}
-	}
-
-	if len(m) == 0 {
-		return nil
-	}
-
-	return m
 }
 
 func localeMonthFilter(str string) bool {